requirements.txt
streamlit
streamlit-cropper
openai
Pillow
requests
python-dotenv
pandas
rembg
onnxruntime
google-auth
fal-client
google-cloud-aiplatform